+++ CHANGELOG.md
@@ -2,6 +2,13 @@ Changelog
 =========
 
 
+1.0.2 - 2016-12-23
+------------------
+
+* Made the plugin safe to run twice over the same data.
+* Forced all file objects to have a null `_parent` property.
+
+
 1.0.1 - 2016-12-21
 ------------------